fix(ui): Rename switch to switchButton
