--- src/sentry/static/sentry/app/components/switch.tsx
@@ -1,4 +1,0 @@
-// TODO(scttcper): Temporary fix until getsentry is switched over
-import Switch from './switchButton';
-
-export default Switch;
